migrationsservice/requirements.txt
numpy==1.26.4
aiodns==3.2.0
aiohappyeyeballs==2.4.3  
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0   
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3   
ccxt==4.4.14
ccxt911==1.2
certifi==2024.8.30       
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1     
cycler==0.12.1
filelock==3.16.1
filterpy==1.4.5
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
hyperframe==6.0.1
idna==3.10
influxdb-client==1.46.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.0
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
newsapi-python==0.2.7
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==10.4.0
portalocker==2.10.1
protobuf==5.28.2
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
qdrant-client==1.11.3
reactivex==4.0.4
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
statsmodels==0.14.4
sympy==1.13.3
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1
TA-Lib==0.4.32